Add files via upload
For the current project, I focused on demonstrating the performance of machine learning models using a binary classification approach, as the dataset contains multiclass labels. To facilitate the analysis, I converted the multiclass labels into binary classes using a one-hot encoder. While this approach provided valuable insights into the model's efficacy and performance, I recognize the importance of showcasing results using the original multiclass labels in future projects. By transitioning to multiclass classification, I aim to offer a more comprehensive understanding of the models' capabilities and their applicability across diverse scenarios. This strategic shift will enable a deeper exploration of the dataset's nuances and enhance the interpretability of the results, laying a foundation for more nuanced and insightful analyses in subsequent projects.